// packages/client/src/api/api.ts
 export interface InlineObject {
     /**
      * 
      * @type {Array<Token>}
      * @memberof InlineObject
      */
     'tokens': Array<Token>;
     /**
      * Ethereum network (Rinkeby / Homestead)
      * @type {string}
      * @memberof InlineObject
      */
     'network': InlineObjectNetworkEnum;
     /**
      * 
      * @type {string}
      * @memberof InlineObject
      */
     'recipient': string;
     /**
      * 
      * @type {string}
      * @memberof InlineObject
      */
     'gnosisSafeAddress': string;
 }
 
 /**
     * @export
     * @enum {string}
     */
 export enum InlineObjectNetworkEnum {
     Rinkeby = 'rinkeby',
     Homestead = 'homestead'
 }
// ...
 export interface Token {
     /**
      * 
      * @type {string}
      * @memberof Token
      */
     'contractAddress': string;
     /**
      * 
      * @type {string}
      * @memberof Token
      */
     'id': string;
 }
// ...
 export class DefaultApi extends BaseAPI {
     /**
      * 
      * @summary Create a batch transaction
      * @param {InlineObject} [body] 
      * @param {*} [options] Override http request option.
      * @throws {RequiredError}
      * @memberof DefaultApi
      */
     public createBatchTransaction(body?: InlineObject, options?: AxiosRequestConfig) {
         return DefaultApiFp(this.configuration).createBatchTransaction(body, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
      * 
      * @summary Fetch details of an OpenSea asset
      * @param {string} contractAddress Contract address for the token to lookup
      * @param {string} id Token ID to lookup
      * @param {'rinkeby' | 'homestead'} network Ethereum network (Rinkeby / Homestead)
      * @param {*} [options] Override http request option.
      * @throws {RequiredError}
      * @memberof DefaultApi
      */
     public getAsset(contractAddress: string, id: string, network: 'rinkeby' | 'homestead', options?: AxiosRequestConfig) {
         return DefaultApiFp(this.configuration).getAsset(contractAddress, id, network, options).then((request) => request(this.axios, this.basePath));
     }
 }